refactor: Moved playlist API responces
